+++ api-samples/rest-api/get-addons-remove_bg-execute-status.php
@@ -1,0 +1,7 @@
+<?php
+
+$configuration = Uploadcare\Configuration::create((string) $_ENV['UPLOADCARE_PUBLIC_KEY'], (string) $_ENV['UPLOADCARE_SECRET_KEY']);
+$api = (new Uploadcare\Api($configuration))->addons();
+$status = $api->checkRemoveBackground('request-id');
+
+echo \sprintf('Remove background status: %s', $status);